+++ lifecycle/lifecycle/endpoints/auth.py
@@ -3,7 +3,7 @@ from fastapi.responses import Response, JSONResponse
 
 from lifecycle.auth.authenticate import authenticate_token
 from lifecycle.auth.authorize import authorize_internal_token, authorize_resource_access, grant_permission
-from lifecycle.auth.subject import get_auth_subject_by_esc, get_auth_subject_by_job_family, regenerate_all_esc_tokens, regenerate_all_job_family_tokens, regenerate_all_user_tokens
+from lifecycle.auth.subject import get_auth_subject_by_esc, get_auth_subject_by_job_family, regenerate_all_esc_tokens, regenerate_all_job_family_tokens, regenerate_all_user_tokens, regenerate_specific_user_token
 from lifecycle.config import Config
 from lifecycle.auth.check import check_auth
 from lifecycle.job.esc import read_esc_model
@@ -81,6 +81,12 @@ def setup_auth_endpoints(api: APIRouter, config: Config):
         return {'token': auth_subject.token}
 
     @api.post('/auth/token/user/regenerate')
+    def _generate_tokens_for_user(request: Request) -> str:
+        """Generate new token for a User"""
+        auth_subject = check_auth(request, subject_types=[AuthSubjectType.USER])
+        return regenerate_specific_user_token(auth_subject)
+
+    @api.post('/auth/token/user/all/regenerate')
     def _generate_tokens_for_all_users(request: Request):
         """Generate new tokens for all Users"""
         check_auth(request, scope=AuthScope.CALL_ADMIN_API)